requirements.txt
SQLAlchemy==2.0.9
jupyterlab==3.6.3
psycopg2-binary==2.9.6
pydantic==1.10.7
uvicorn==0.22.0
Jinja2==3.0.3
fastapi==0.73.0
python-jose==3.3.0
itsdangerous==2.1.2
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
